+++ CHANGELOG.md
@@ -3,6 +3,12 @@
 All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## 1.9.7
+
+November 22, 2023
+
+New version of subdaily, datetime on most x-axis plots, removes gaps in the txt file.
+
 ## 1.9.6
 November 18, 2023